+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # api-session-ts Changelog
 
+## 0.0.2 (unreleased)
+
+* Fix `baseUrl`, so you can import from `api-session-ts` instead of `api-session-ts/src`
+* Add `HTTPError`
+* Add default headers, default user-agent
+* Add `.fetch()`, `.getJSON()`, `postJSON`, `putJSON`, `get`, `head`, `delete_`, `options`
+
 ## 0.0.1 (2025/01/21)
 
 Initial release (WIP).